requirements.txt
PyPDF2==3.0.1
spacy==3.5.3
transformers==4.29.2
torch==2.0.1
pandas==2.0.2
scikit-learn==1.2.2
tqdm==4.65.0
pyyaml==6.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.27.8
flask==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.9
aiohttp

